Update Rivet API and Devtools Dock
This commit updates the Rivet API and Devtools Dock. It includes changes to the `_get_api_url()` function in `rivet_api.gd` to use `RivetDevtools.get_plugin().api_endpoint` instead of a hardcoded URL. It also adds a new `ApiEndpointTextEdit` in `login.gd` to allow the user to set the API endpoint. Additionally, it fixes a typo in the `settings.gd` file.